bot/requirements.txt
discord.py==2.3.2
fastapi==0.115.0
uvicorn==0.30.3
